Read (Load) SomaLogic ADATs

The parse and load a *.adat file as a data.frame-like object into an R workspace environment. The class of the returned object is a soma_adat object.

read.adat() is a convenient alias for read_adat() designed to enable backward compatibility to older versions of SomaDataIO.

is.soma_adat() checks whether an object is of class soma_adat. See inherits().

Subsetting a soma_adat object class relies heavily on the S3 ([) method of the generic Extract() function that keeps the soma_adat attributes intact and subsets the Col.Meta so that it is consistent with the new object.

read_adat(file, debug = FALSE, verbose = getOption("verbose"), ...)

read.adat(file, debug = FALSE, verbose = getOption("verbose"), ...)

is.soma_adat(x)

# S3 method for soma_adat
[(x, i, j, drop = TRUE)

# S3 method for soma_adat
print(x, show_header = FALSE, ...)

# S3 method for soma_adat
summary(object, tbl, digits = max(3L, getOption("digits") - 3L), ...)

Arguments

file

Character. The elaborated path and file name of the *.adat file to be loaded into an R workspace.

debug

Logical. Used for debugging and development of an ADAT that fails to load, particularly out-of-spec, poorly modified, or legacy ADATs.

verbose

Logical. Should the function call be run in verbose mode, printing relevant diagnostic call information to the console.

...

Additional arguments passed ultimately to read_delim(), or additional arguments passed to either other S3 print or summary methods as required by those generics.

x

A soma_adat class object to print().

i

indices specifying elements to extract or replace. Indices are numeric or character vectors or empty (missing) or NULL. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero). Character vectors will be matched to the names of the object (or for matrices/arrays, the dimnames): see ‘Character indices’ below for further details.

For [-indexing only: i, j, ... can be logical vectors, indicating elements/slices to select. Such vectors are recycled if necessary to match the corresponding extent. i, j, ... can also be negative integers, indicating elements/slices to leave out of the selection.

When indexing arrays by [ a single argument i can be a matrix with as many columns as there are dimensions of x; the result is then a vector with elements corresponding to the sets of indices in each row of i.

An index value of NULL is treated as if it were integer(0).

j

indices specifying elements to extract or replace. Indices are numeric or character vectors or empty (missing) or NULL. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero). Character vectors will be matched to the names of the object (or for matrices/arrays, the dimnames): see ‘Character indices’ below for further details.

For [-indexing only: i, j, ... can be logical vectors, indicating elements/slices to select. Such vectors are recycled if necessary to match the corresponding extent. i, j, ... can also be negative integers, indicating elements/slices to leave out of the selection.

When indexing arrays by [ a single argument i can be a matrix with as many columns as there are dimensions of x; the result is then a vector with elements corresponding to the sets of indices in each row of i.

An index value of NULL is treated as if it were integer(0).

drop

For matrices and arrays. If TRUE the result is coerced to the lowest possible dimension (see the examples). This only works for extracting elements, not for the replacement. See drop for further details.

show_header

Logical. Should all the Header Data information be displayed instead of the data frame (tibble) object?

object

Object of class soma_adat used in the S3 generic summary().

tbl

An annotations table. If NULL (default), annotation information is extracted from the object itself (if possible). Alternatively, the result of a call to getAnalyteInfo(), from which Target names can be extracted.

digits

Integer. Used for number formatting with signif().

Value

A data.frame-like object of class soma_adatconsisting of SomaLogic RFU (feature) data and clinical meta data as columns, and samples as rows. Row names are labeled with the unique ID "SlideId_Subarray" concatenation. The sections of the ADAT header (e.g., "Header.Meta", "Col.Meta", ...) are stored as attributes (e.g. attributes(x)$Header.Meta). A logical indicating whether x inherits from class soma_adat.

Methods (by generic)

print: The S3 generic print method (print) returns summary information parsed from the object attributes, if possible (see examples), followed by a dispatch to the tibble print method.

summary: An associated S3 generic summary method (summary()) returns the following for each column of the ADAT object containing SOMAmer data (clinical meta data is excluded):

Target (if available)

Minimum value

1st Quantile

Median

Mean

3rd Quantile

Maximum value

Standard deviation

Median absolute deviation (mad())

Interquartile range (IQR())

See also

Author

Stu Field

Examples

f <- system.file("example", "example_data.adat",
                 package = "SomaDataIO", mustWork = TRUE)
my_adat <- read_adat(f)
is.soma_adat(my_adat)
#> [1] TRUE

# S3 print method
my_adat                             # redirect uses `tibble` method
#> ══ SomaScan Data ═══════════════════════════════════════════════════════════════
#>      Attributes intact    ✓
#>      Rows                 192
#>      Columns              5318
#>      Clinical Data        34
#>      Features             5284
#> ── Column Meta ─────────────────────────────────────────────────────────────────
#> ℹ SeqId, SeqIdVersion, SomaId, TargetFullName, Target, UniProt,
#> ℹ EntrezGeneID, EntrezGeneSymbol, Organism, Units, Type, Dilution,
#> ℹ PlateScale_Reference, CalReference, Cal_Example_Adat_Set001,
#> ℹ ColCheck, CalQcRatio_Example_Adat_Set001_170255, QcReference_170255,
#> ℹ Cal_Example_Adat_Set002, CalQcRatio_Example_Adat_Set002_170255,
#> ℹ Dilution2
#> ── Tibble ──────────────────────────────────────────────────────────────────────
#> # A tibble: 192 × 5,319
#>    row_names      PlateId  PlateRunDate ScannerID PlatePosition SlideId Subarray
#>    <chr>          <chr>    <date>       <chr>     <chr>           <dbl>    <dbl>
#>  1 258495800012_3 Example… 2020-06-18   SG152144… H9            2.58e11        3
#>  2 258495800004_7 Example… 2020-06-18   SG152144… H8            2.58e11        7
#>  3 258495800010_8 Example… 2020-06-18   SG152144… H7            2.58e11        8
#>  4 258495800003_4 Example… 2020-06-18   SG152144… H6            2.58e11        4
#>  5 258495800009_4 Example… 2020-06-18   SG152144… H5            2.58e11        4
#>  6 258495800012_8 Example… 2020-06-18   SG152144… H4            2.58e11        8
#>  7 258495800001_3 Example… 2020-06-18   SG152144… H3            2.58e11        3
#>  8 258495800004_8 Example… 2020-06-18   SG152144… H2            2.58e11        8
#>  9 258495800001_8 Example… 2020-06-18   SG152144… H12           2.58e11        8
#> 10 258495800004_3 Example… 2020-06-18   SG152144… H11           2.58e11        3
#> # … with 182 more rows, and 5,312 more variables: SampleId <dbl>,
#> #   SampleType <chr>, PercentDilution <dbl>, SampleMatrix <chr>, Barcode <lgl>,
#> #   Barcode2d <lgl>, SampleName <lgl>, SampleNotes <lgl>,
#> #   AliquotingNotes <lgl>, SampleDescription <chr>, AssayNotes <lgl>,
#> #   TimePoint <lgl>, ExtIdentifier <lgl>, SsfExtId <lgl>, SampleGroup <lgl>, …
#> ════════════════════════════════════════════════════════════════════════════════
print(my_adat, show_header = TRUE)  # show the header info; no RFU data
#> ══ SomaScan Data ═══════════════════════════════════════════════════════════════
#>      Attributes intact    ✓
#>      Rows                 192
#>      Columns              5318
#>      Clinical Data        34
#>      Features             5284
#> ── Column Meta ─────────────────────────────────────────────────────────────────
#> ℹ SeqId, SeqIdVersion, SomaId, TargetFullName, Target, UniProt,
#> ℹ EntrezGeneID, EntrezGeneSymbol, Organism, Units, Type, Dilution,
#> ℹ PlateScale_Reference, CalReference, Cal_Example_Adat_Set001,
#> ℹ ColCheck, CalQcRatio_Example_Adat_Set001_170255, QcReference_170255,
#> ℹ Cal_Example_Adat_Set002, CalQcRatio_Example_Adat_Set002_170255,
#> ℹ Dilution2
#> ── Header Data ─────────────────────────────────────────────────────────────────
#> # A tibble: 35 × 2
#>    Key                  Value                                                   
#>    <chr>                <chr>                                                   
#>  1 AdatId               GID-1234-56-789-abcdef                                  
#>  2 Version              1.2                                                     
#>  3 AssayType            PharmaServices                                          
#>  4 AssayVersion         V4                                                      
#>  5 AssayRobot           Fluent 1 L-307                                          
#>  6 Legal                Experiment details and data have been processed to prot…
#>  7 CreatedBy            PharmaServices                                          
#>  8 CreatedDate          2020-07-24                                              
#>  9 EnteredBy            Technician1                                             
#> 10 ExpDate              2020-06-18, 2020-07-20                                  
#> 11 GeneratedBy          Px (Build:  : ), Canopy_0.1.1                           
#> 12 RunNotes             2 columns ('Age' and 'Sex') have been added to this ADA…
#> 13 ProcessSteps         Raw RFU, Hyb Normalization, medNormInt (SampleId), plat…
#> 14 ProteinEffectiveDate 2019-08-06                                              
#> 15 StudyMatrix          EDTA Plasma                                             
#> # … with 20 more rows
#> ════════════════════════════════════════════════════════════════════════════════

# write ADAT to file
fout <- tempfile(fileext = ".adat")
write_adat(my_adat, file = fout)
#> ✔ ADAT passed checks and traps
#> ✔ ADAT written to: '/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpB27ZQk/file125064c50979.adat'

# read same file back in as check
read_adat(fout)
#> ══ SomaScan Data ═══════════════════════════════════════════════════════════════
#>      Attributes intact    ✓
#>      Rows                 192
#>      Columns              5318
#>      Clinical Data        34
#>      Features             5284
#> ── Column Meta ─────────────────────────────────────────────────────────────────
#> ℹ SeqId, SeqIdVersion, SomaId, TargetFullName, Target, UniProt,
#> ℹ EntrezGeneID, EntrezGeneSymbol, Organism, Units, Type, Dilution,
#> ℹ PlateScale_Reference, CalReference, Cal_Example_Adat_Set001,
#> ℹ ColCheck, CalQcRatio_Example_Adat_Set001_170255, QcReference_170255,
#> ℹ Cal_Example_Adat_Set002, CalQcRatio_Example_Adat_Set002_170255,
#> ℹ Dilution2
#> ── Tibble ──────────────────────────────────────────────────────────────────────
#> # A tibble: 192 × 5,319
#>    row_names      PlateId  PlateRunDate ScannerID PlatePosition SlideId Subarray
#>    <chr>          <chr>    <date>       <chr>     <chr>           <dbl>    <dbl>
#>  1 258495800012_3 Example… 2020-06-18   SG152144… H9            2.58e11        3
#>  2 258495800004_7 Example… 2020-06-18   SG152144… H8            2.58e11        7
#>  3 258495800010_8 Example… 2020-06-18   SG152144… H7            2.58e11        8
#>  4 258495800003_4 Example… 2020-06-18   SG152144… H6            2.58e11        4
#>  5 258495800009_4 Example… 2020-06-18   SG152144… H5            2.58e11        4
#>  6 258495800012_8 Example… 2020-06-18   SG152144… H4            2.58e11        8
#>  7 258495800001_3 Example… 2020-06-18   SG152144… H3            2.58e11        3
#>  8 258495800004_8 Example… 2020-06-18   SG152144… H2            2.58e11        8
#>  9 258495800001_8 Example… 2020-06-18   SG152144… H12           2.58e11        8
#> 10 258495800004_3 Example… 2020-06-18   SG152144… H11           2.58e11        3
#> # … with 182 more rows, and 5,312 more variables: SampleId <dbl>,
#> #   SampleType <chr>, PercentDilution <dbl>, SampleMatrix <chr>, Barcode <lgl>,
#> #   Barcode2d <lgl>, SampleName <lgl>, SampleNotes <lgl>,
#> #   AliquotingNotes <lgl>, SampleDescription <chr>, AssayNotes <lgl>,
#> #   TimePoint <lgl>, ExtIdentifier <lgl>, SsfExtId <lgl>, SampleGroup <lgl>, …
#> ════════════════════════════════════════════════════════════════════════════════

# S3 summary method
# MMP (4) analytes
mmps <- c("seq.2579.17", "seq.2788.55", "seq.2789.26", "seq.4925.54")
summary(my_adat[, mmps])
#>  seq.2579.17         seq.2788.55         seq.2789.26         
#>  Target : MMP-9      Target : MMP-3      Target : MMP-7      
#>  Min    :    59.3    Min    :  29.30     Min    :    24.1    
#>  1Q     :  9360.5    1Q     : 126.67     1Q     :  8182.0    
#>  Median : 13191.2    Median : 149.75     Median : 10947.8    
#>  Mean   : 14372.5    Mean   : 157.78     Mean   : 11277.6    
#>  3Q     : 19387.3    3Q     : 179.08     3Q     : 13577.0    
#>  Max    : 51263.0    Max    : 423.20     Max    : 33496.9    
#>  sd     :  7449.9    sd     :  53.57     sd     :  5802.6    
#>  MAD    :  6877.9    MAD    :  38.62     MAD    :  4001.2    
#>  IQR    : 10026.8    IQR    :  52.40     IQR    :  5395.0    
#>  seq.4925.54         
#>  Target : MMP-13     
#>  Min    :   25.20    
#>  1Q     :  328.30    
#>  Median :  389.70    
#>  Mean   :  424.48    
#>  3Q     :  452.85    
#>  Max    : 2883.20    
#>  sd     :  259.01    
#>  MAD    :   93.85    
#>  IQR    :  124.55    

# Summarize by group
my_adat[, mmps] %>%
  split(my_adat$Sex) %>%
  lapply(summary)
#> $F
#>  seq.2579.17         seq.2788.55         seq.2789.26         
#>  Target : MMP-9      Target : MMP-3      Target : MMP-7      
#>  Min    :  4484      Min    :  97.10     Min    :   289.7    
#>  1Q     :  8336      1Q     : 127.70     1Q     :  9410.0    
#>  Median : 12859      Median : 150.60     Median : 12157.2    
#>  Mean   : 13674      Mean   : 156.02     Mean   : 12877.3    
#>  3Q     : 16148      3Q     : 181.10     3Q     : 15151.9    
#>  Max    : 31879      Max    : 235.40     Max    : 32221.3    
#>  sd     :  6389      sd     :  34.66     sd     :  5214.4    
#>  MAD    :  6488      MAD    :  38.10     MAD    :  4342.7    
#>  IQR    :  7811      IQR    :  53.40     IQR    :  5741.9    
#>  seq.4925.54         
#>  Target : MMP-13     
#>  Min    :  241.80    
#>  1Q     :  333.80    
#>  Median :  388.10    
#>  Mean   :  422.14    
#>  3Q     :  460.70    
#>  Max    : 1353.80    
#>  sd     :  171.89    
#>  MAD    :   95.78    
#>  IQR    :  126.90    
#> 
#> $M
#>  seq.2579.17         seq.2788.55         seq.2789.26         
#>  Target : MMP-9      Target : MMP-3      Target : MMP-7      
#>  Min    :  5894      Min    :  99.90     Min    :  2850      
#>  1Q     :  9973      1Q     : 132.40     1Q     :  8640      
#>  Median : 13645      Median : 157.40     Median : 11230      
#>  Mean   : 15715      Mean   : 173.63     Mean   : 11699      
#>  3Q     : 20595      3Q     : 196.20     3Q     : 13061      
#>  Max    : 51263      Max    : 423.20     Max    : 33497      
#>  sd     :  7979      sd     :  61.17     sd     :  5243      
#>  MAD    :  6344      MAD    :  40.18     MAD    :  3477      
#>  IQR    : 10622      IQR    :  63.80     IQR    :  4421      
#>  seq.4925.54         
#>  Target : MMP-13     
#>  Min    :  205.7     
#>  1Q     :  329.3     
#>  Median :  401.6     
#>  Mean   :  462.7     
#>  3Q     :  468.9     
#>  Max    : 2883.2     
#>  sd     :  331.3     
#>  MAD    :  103.8     
#>  IQR    :  139.6     
#> 

# Alternatively pass annotations with Target info
anno <- getAnalyteInfo(my_adat)
summary(my_adat[, mmps], tbl = anno)
#>  seq.2579.17         seq.2788.55         seq.2789.26         
#>  Target : MMP-9      Target : MMP-3      Target : MMP-7      
#>  Min    :    59.3    Min    :  29.30     Min    :    24.1    
#>  1Q     :  9360.5    1Q     : 126.67     1Q     :  8182.0    
#>  Median : 13191.2    Median : 149.75     Median : 10947.8    
#>  Mean   : 14372.5    Mean   : 157.78     Mean   : 11277.6    
#>  3Q     : 19387.3    3Q     : 179.08     3Q     : 13577.0    
#>  Max    : 51263.0    Max    : 423.20     Max    : 33496.9    
#>  sd     :  7449.9    sd     :  53.57     sd     :  5802.6    
#>  MAD    :  6877.9    MAD    :  38.62     MAD    :  4001.2    
#>  IQR    : 10026.8    IQR    :  52.40     IQR    :  5395.0    
#>  seq.4925.54         
#>  Target : MMP-13     
#>  Min    :   25.20    
#>  1Q     :  328.30    
#>  Median :  389.70    
#>  Mean   :  424.48    
#>  3Q     :  452.85    
#>  Max    : 2883.20    
#>  sd     :  259.01    
#>  MAD    :   93.85    
#>  IQR    :  124.55