Write an ADAT to File

One can write an existing modified internal ADAT (soma_adat R object) to an external file. However the ADAT object itself must have intact attributes, see is.intact.attributes().

write_adat(x, file)

Arguments

x

An object of class "soma_adat". Both is.soma_adat() and is.intact.attributes() must be TRUE.

file

Character. File path where the object should be written. For example, extensions should be *.adat.

Details

The ADAT specification no longer requires Windows end of line (EOL) characters ("\r\n"). The current EOL spec is "\n" which is commonly used in POSIX systems, like MacOS and Linux. Since the EOL affects the resulting checksum, ADATs written on other systems generate slightly differing files. Standardizing to "\n" attempts to solve this issue. For reference, see the EOL encoding for operating systems below:

Symbol	Platform	Character

LF	Linux	"\n"

CR	MacOS	"\r"

CRLF	DOS/Windows	"\r\n"

See also

Author

Stu Field