SomaDataIO: SomaData Input/Output Package

The SomaDataIO package loads the SomaLogic Operating Co., Inc. proprietary data file, called an ADAT file ('*.adat'). It also provides auxiliary functions for extracting and manipulating relevant information from an ADAT, as well as exporting a modified ADAT to file.

Details

Load an ADAT file into the global workspace with a call to read_adat(). This function parses the main data table into a data.frame object and assigns the remaining data from the file as object attributes, i.e. call attributes(adat). Other functions in the package are designed to make extracting, manipulating, and wrangling data in the newly created soma_adat object more convenient.

Those familiar with micro-array data analysis and associated packages, e.g. Biobase, will notice that the feature data (proteins) are arranged as columns and the samples (arrays) are the rows. This is the transpose of typical micro-array data. This conflict can be easily solved using the transpose function, t(), which is part of the base R. In addition, those familiar with the standard ExpressionSet object, available from Bioconductor, might find the functions adat2eSet() and pivotExpressionSet() particularly useful.

See also

Author

Maintainer: Stu Field sfield@somalogic.com

Other contributors:

SomaLogic Operating Co., Inc. [copyright holder, funder]

Examples

# For a listing of all SomaDataIO functions
library(help = SomaDataIO)

# To load an original adat call read_adat():
# There is a sample *.adat file provided in SomaDataIO
file <- system.file("example", "example_data.adat",
                    package = "SomaDataIO", mustWork = TRUE)
my_adat <- read_adat(file)

# Object class `soma_adat`
class(my_adat)
#> [1] "soma_adat"  "data.frame"
is.soma_adat(my_adat)
#> [1] TRUE

# Annotations Lookup Table
anno_tbl <- getAnalyteInfo(my_adat)
anno_tbl
#> # A tibble: 5,284 × 22
#>    AptName  SeqId SeqIdVersion SomaId TargetFullName Target UniProt EntrezGeneID
#>    <chr>    <chr>        <dbl> <chr>  <chr>          <chr>  <chr>   <chr>       
#>  1 seq.100… 1000…            3 SL019… Beta-crystall… CRBB2  P43320  "1415"      
#>  2 seq.100… 1000…            3 SL002… RAF proto-onc… c-Raf  P04049  "5894"      
#>  3 seq.100… 1000…            3 SL019… Zinc finger p… ZNF41  P51814  "7592"      
#>  4 seq.100… 1000…            3 SL019… ETS domain-co… ELK1   P19419  "2002"      
#>  5 seq.100… 1000…            3 SL019… Guanylyl cycl… GUC1A  P43080  "2978"      
#>  6 seq.100… 1001…            3 SL019… Inositol poly… OCRL   Q01968  "4952"      
#>  7 seq.100… 1001…            3 SL014… SAM pointed d… SPDEF  O95238  "25803"     
#>  8 seq.100… 1001…            3 SL025… Fc_MOUSE       Fc_MO… Q99LC4  ""          
#>  9 seq.100… 1001…            3 SL007… Zinc finger p… SLUG   O43623  "6591"      
#> 10 seq.100… 1001…            3 SL014… Voltage-gated… KCAB2  Q13303  "8514"      
#> # … with 5,274 more rows, and 14 more variables: EntrezGeneSymbol <chr>,
#> #   Organism <chr>, Units <chr>, Type <chr>, Dilution <chr>,
#> #   PlateScale_Reference <dbl>, CalReference <dbl>,
#> #   Cal_Example_Adat_Set001 <dbl>, ColCheck <chr>,
#> #   CalQcRatio_Example_Adat_Set001_170255 <dbl>, QcReference_170255 <dbl>,
#> #   Cal_Example_Adat_Set002 <dbl>, CalQcRatio_Example_Adat_Set002_170255 <dbl>,
#> #   Dilution2 <dbl>

# Find all analytes starting with "MMP" in `anno_tbl`
anno_tbl %>% dplyr::filter(grepl("^MMP", Target))
#> # A tibble: 15 × 22
#>    AptName  SeqId SeqIdVersion SomaId TargetFullName Target UniProt EntrezGeneID
#>    <chr>    <chr>        <dbl> <chr>  <chr>          <chr>  <chr>   <chr>       
#>  1 seq.154… 1541…            3 SL012… Matrix metall… MMP20  O60882  9313        
#>  2 seq.257… 2579…            5 SL000… Matrix metall… MMP-9  P14780  4318        
#>  3 seq.278… 2788…            1 SL000… Stromelysin-1  MMP-3  P08254  4314        
#>  4 seq.278… 2789…            2 SL000… Matrilysin     MMP-7  P09237  4316        
#>  5 seq.283… 2838…            1 SL003… Matrix metall… MMP-17 Q9ULZ9  4326        
#>  6 seq.416… 4160…            1 SL000… 72 kDa type I… MMP-2  P08253  4313        
#>  7 seq.449… 4496…            2 SL000… Macrophage me… MMP-12 P39900  4321        
#>  8 seq.492… 4924…            1 SL000… Interstitial … MMP-1  P03956  4312        
#>  9 seq.492… 4925…            2 SL000… Collagenase 3  MMP-13 P45452  4322        
#> 10 seq.500… 5002…            1 SL002… Matrix metall… MMP-14 P50281  4323        
#> 11 seq.526… 5268…            3 SL003… Matrix metall… MMP-16 P51512  4325        
#> 12 seq.642… 6425…            3 SL007… Matrix metall… MMP19  Q99542  4327        
#> 13 seq.847… 8479…            3 SL000… Stromelysin-2  MMP-10 P09238  4319        
#> 14 seq.917… 9172…            3 SL000… Neutrophil co… MMP-8  P22894  4317        
#> 15 seq.971… 9719…            3 SL003… Matrix metall… MMP-16 P51512  4325        
#> # … with 14 more variables: EntrezGeneSymbol <chr>, Organism <chr>,
#> #   Units <chr>, Type <chr>, Dilution <chr>, PlateScale_Reference <dbl>,
#> #   CalReference <dbl>, Cal_Example_Adat_Set001 <dbl>, ColCheck <chr>,
#> #   CalQcRatio_Example_Adat_Set001_170255 <dbl>, QcReference_170255 <dbl>,
#> #   Cal_Example_Adat_Set002 <dbl>, CalQcRatio_Example_Adat_Set002_170255 <dbl>,
#> #   Dilution2 <dbl>