SomaDataIO: Input/Output 'SomaScan' Data

Load and export 'SomaScan' data via the 'SomaLogic Operating Co., Inc.' proprietary text file called an ADAT ('*.adat'). For file format see https://github.com/SomaLogic/SomaLogic-Data/blob/master/README.md. The package also exports auxiliary functions for manipulating, wrangling, and extracting relevant information from an ADAT object once in memory.

Details

Load an ADAT file into the global workspace with a call to read_adat(). This function parses the main data table into a data.frame object and assigns the remaining data from the file as object attributes, i.e. call attributes(adat). Other functions in the package are designed to make extracting, manipulating, and wrangling data in the newly created soma_adat object more convenient.

Those familiar with micro-array data analysis and associated packages, e.g. Biobase, will notice that the feature data (proteins) are arranged as columns and the samples (arrays) are the rows. This is the transpose of typical micro-array data. This conflict can be easily solved using the transpose function, t(), which is part of the base R. In addition, those familiar with the standard ExpressionSet object, available from Bioconductor, might find the functions adat2eSet() and pivotExpressionSet() particularly useful.

See also

Author

Maintainer: Stu Field stu.g.field@gmail.com (ORCID)

Other contributors:

SomaLogic Operating Co., Inc. [copyright holder, funder]

Examples

# a listing of all pkg functions
library(help = SomaDataIO)

# the `soma_adat` class
class(example_data)
#> [1] "soma_adat"  "data.frame"
is.soma_adat(example_data)
#> [1] TRUE

# Annotations Lookup Table
anno_tbl <- getAnalyteInfo(example_data)
anno_tbl
#> # A tibble: 5,284 × 22
#>    AptName   SeqId SeqId…¹ SomaId Targe…² Target UniProt Entre…³ Entre…⁴ Organ…⁵
#>    <chr>     <chr>   <dbl> <chr>  <chr>   <chr>  <chr>   <chr>   <chr>   <chr>  
#>  1 seq.1000… 1000…       3 SL019… Beta-c… CRBB2  P43320  "1415"  "CRYBB… Human  
#>  2 seq.1000… 1000…       3 SL002… RAF pr… c-Raf  P04049  "5894"  "RAF1"  Human  
#>  3 seq.1000… 1000…       3 SL019… Zinc f… ZNF41  P51814  "7592"  "ZNF41" Human  
#>  4 seq.1000… 1000…       3 SL019… ETS do… ELK1   P19419  "2002"  "ELK1"  Human  
#>  5 seq.1000… 1000…       3 SL019… Guanyl… GUC1A  P43080  "2978"  "GUCA1… Human  
#>  6 seq.1001… 1001…       3 SL019… Inosit… OCRL   Q01968  "4952"  "OCRL"  Human  
#>  7 seq.1001… 1001…       3 SL014… SAM po… SPDEF  O95238  "25803" "SPDEF" Human  
#>  8 seq.1001… 1001…       3 SL025… Fc_MOU… Fc_MO… Q99LC4  ""      ""      Mouse  
#>  9 seq.1001… 1001…       3 SL007… Zinc f… SLUG   O43623  "6591"  "SNAI2" Human  
#> 10 seq.1001… 1001…       3 SL014… Voltag… KCAB2  Q13303  "8514"  "KCNAB… Human  
#> # … with 5,274 more rows, 12 more variables: Units <chr>, Type <chr>,
#> #   Dilution <chr>, PlateScale_Reference <dbl>, CalReference <dbl>,
#> #   Cal_Example_Adat_Set001 <dbl>, ColCheck <chr>,
#> #   CalQcRatio_Example_Adat_Set001_170255 <dbl>, QcReference_170255 <dbl>,
#> #   Cal_Example_Adat_Set002 <dbl>, CalQcRatio_Example_Adat_Set002_170255 <dbl>,
#> #   Dilution2 <dbl>, and abbreviated variable names ¹​SeqIdVersion,
#> #   ²​TargetFullName, ³​EntrezGeneID, ⁴​EntrezGeneSymbol, ⁵​Organism

# Find all analytes starting with "MMP" in `anno_tbl`
dplyr::filter(anno_tbl, grepl("^MMP", Target))
#> # A tibble: 15 × 22
#>    AptName   SeqId SeqId…¹ SomaId Targe…² Target UniProt Entre…³ Entre…⁴ Organ…⁵
#>    <chr>     <chr>   <dbl> <chr>  <chr>   <chr>  <chr>   <chr>   <chr>   <chr>  
#>  1 seq.1541… 1541…       3 SL012… Matrix… MMP20  O60882  9313    MMP20   Human  
#>  2 seq.2579… 2579…       5 SL000… Matrix… MMP-9  P14780  4318    MMP9    Human  
#>  3 seq.2788… 2788…       1 SL000… Strome… MMP-3  P08254  4314    MMP3    Human  
#>  4 seq.2789… 2789…       2 SL000… Matril… MMP-7  P09237  4316    MMP7    Human  
#>  5 seq.2838… 2838…       1 SL003… Matrix… MMP-17 Q9ULZ9  4326    MMP17   Human  
#>  6 seq.4160… 4160…       1 SL000… 72 kDa… MMP-2  P08253  4313    MMP2    Human  
#>  7 seq.4496… 4496…       2 SL000… Macrop… MMP-12 P39900  4321    MMP12   Human  
#>  8 seq.4924… 4924…       1 SL000… Inters… MMP-1  P03956  4312    MMP1    Human  
#>  9 seq.4925… 4925…       2 SL000… Collag… MMP-13 P45452  4322    MMP13   Human  
#> 10 seq.5002… 5002…       1 SL002… Matrix… MMP-14 P50281  4323    MMP14   Human  
#> 11 seq.5268… 5268…       3 SL003… Matrix… MMP-16 P51512  4325    MMP16   Human  
#> 12 seq.6425… 6425…       3 SL007… Matrix… MMP19  Q99542  4327    MMP19   Human  
#> 13 seq.8479… 8479…       3 SL000… Strome… MMP-10 P09238  4319    MMP10   Human  
#> 14 seq.9172… 9172…       3 SL000… Neutro… MMP-8  P22894  4317    MMP8    Human  
#> 15 seq.9719… 9719…       3 SL003… Matrix… MMP-16 P51512  4325    MMP16   Human  
#> # … with 12 more variables: Units <chr>, Type <chr>, Dilution <chr>,
#> #   PlateScale_Reference <dbl>, CalReference <dbl>,
#> #   Cal_Example_Adat_Set001 <dbl>, ColCheck <chr>,
#> #   CalQcRatio_Example_Adat_Set001_170255 <dbl>, QcReference_170255 <dbl>,
#> #   Cal_Example_Adat_Set002 <dbl>, CalQcRatio_Example_Adat_Set002_170255 <dbl>,
#> #   Dilution2 <dbl>, and abbreviated variable names ¹​SeqIdVersion,
#> #   ²​TargetFullName, ³​EntrezGeneID, ⁴​EntrezGeneSymbol, ⁵​Organism