Example Data and Objects

The example_data object is intended to provide existing and prospective SomaLogic customers with example data to enable analysis preparation prior to receipt of SomaScan data, and also for those generally curious about the SomaScan data deliverable. It is not intended to be used as a control group for studies or provide any metrics for SomaScan data in general.

Format

example_data

a soma_adat parsed via read_adat() containing 192 samples (see below for breakdown of sample type). There are 5318 columns containing 5284 analyte features and 34 clinical meta data fields. These data have been pre-processed via the following steps:

hybridization normalized (all samples)

calibrators and buffers median normalized

plate scaled

calibrated

Adaptive Normalization by Maximum Likelihood (ANML) of QC and clinical samples

Note1: The Age and Sex (M/F) fields contain simulated values designed to contain biological signal.

**Note2:** The `SampleType` column contains sample source/type information
and usually the `SampleType == Sample` represents the "client" samples.

**Note3:** The original source file can be found at
\url{https://github.com/SomaLogic/SomaLogic-Data}.

ex_analytes

character string of the analyte features contained in the soma_adat object, derived from a call to getAnalytes().

ex_anno_tbl

a lookup table corresponding to a transposed data frame of the "Col.Meta" attribute of an ADAT, with an index key field AptName included in column 1, derived from a call to getAnalyteInfo().

ex_target_names

A lookup table mapping SeqId feature names -> target names contained in example_data. This object (or one like it) is convenient at the console via auto-complete for labeling and/or creating plot titles on the fly.

Source

https://github.com/SomaLogic/SomaLogic-Data

SomaLogic Operating Co., Inc.

Data Description

The example_data object contains a SomaScan V4 study from healthy normal individuals. The RFU measurements themselves and other identifiers have been altered to protect personally identifiable information (PII), but also retain underlying biological signal as much as possible. There are 192 total EDTA-plasma samples across two 96-well plate runs which are broken down by the following types:

170 clinical samples (client study samples)

10 calibrators (replicate controls for combining data across runs)

6 QC samples (replicate controls used to assess run quality)

6 Buffer samples (no protein controls)

Data Processing

The standard V4 data normalization procedure for EDTA-plasma samples was applied to this dataset. For more details on the data standardization process see the Data Standardization and File Specification Technical Note. General details are outlined above.

Examples

# S3 print method
example_data
#> ══ SomaScan Data ═══════════════════════════════════════════════════════════════
#>      Attributes intact    ✓
#>      Rows                 192
#>      Columns              5318
#>      Clinical Data        34
#>      Features             5284
#> ── Column Meta ─────────────────────────────────────────────────────────────────
#> ℹ SeqId, SeqIdVersion, SomaId, TargetFullName, Target, UniProt,
#> ℹ EntrezGeneID, EntrezGeneSymbol, Organism, Units, Type, Dilution,
#> ℹ PlateScale_Reference, CalReference, Cal_Example_Adat_Set001,
#> ℹ ColCheck, CalQcRatio_Example_Adat_Set001_170255, QcReference_170255,
#> ℹ Cal_Example_Adat_Set002, CalQcRatio_Example_Adat_Set002_170255,
#> ℹ Dilution2
#> ── Tibble ──────────────────────────────────────────────────────────────────────
#> # A tibble: 192 × 5,319
#>    row_names     PlateId Plate…¹ Scann…² Plate…³ SlideId Subar…⁴ Sampl…⁵ Sampl…⁶
#>    <chr>         <chr>   <chr>   <chr>   <chr>     <dbl>   <dbl> <chr>   <chr>  
#>  1 258495800012… Exampl… 2020-0… SG1521… H9      2.58e11       3 1       Sample 
#>  2 258495800004… Exampl… 2020-0… SG1521… H8      2.58e11       7 2       Sample 
#>  3 258495800010… Exampl… 2020-0… SG1521… H7      2.58e11       8 3       Sample 
#>  4 258495800003… Exampl… 2020-0… SG1521… H6      2.58e11       4 4       Sample 
#>  5 258495800009… Exampl… 2020-0… SG1521… H5      2.58e11       4 5       Sample 
#>  6 258495800012… Exampl… 2020-0… SG1521… H4      2.58e11       8 6       Sample 
#>  7 258495800001… Exampl… 2020-0… SG1521… H3      2.58e11       3 7       Sample 
#>  8 258495800004… Exampl… 2020-0… SG1521… H2      2.58e11       8 8       Sample 
#>  9 258495800001… Exampl… 2020-0… SG1521… H12     2.58e11       8 9       Sample 
#> 10 258495800004… Exampl… 2020-0… SG1521… H11     2.58e11       3 170261  Calibr…
#> # … with 182 more rows, 5,310 more variables: PercentDilution <int>,
#> #   SampleMatrix <chr>, Barcode <lgl>, Barcode2d <chr>, SampleName <lgl>,
#> #   SampleNotes <lgl>, AliquotingNotes <lgl>, SampleDescription <chr>,
#> #   AssayNotes <lgl>, TimePoint <lgl>, …, and abbreviated variable names
#> #   ¹​PlateRunDate, ²​ScannerID, ³​PlatePosition, ⁴​Subarray, ⁵​SampleId,
#> #   ⁶​SampleType
#> ════════════════════════════════════════════════════════════════════════════════

# print header info
print(example_data, show_header = TRUE)
#> ══ SomaScan Data ═══════════════════════════════════════════════════════════════
#>      Attributes intact    ✓
#>      Rows                 192
#>      Columns              5318
#>      Clinical Data        34
#>      Features             5284
#> ── Column Meta ─────────────────────────────────────────────────────────────────
#> ℹ SeqId, SeqIdVersion, SomaId, TargetFullName, Target, UniProt,
#> ℹ EntrezGeneID, EntrezGeneSymbol, Organism, Units, Type, Dilution,
#> ℹ PlateScale_Reference, CalReference, Cal_Example_Adat_Set001,
#> ℹ ColCheck, CalQcRatio_Example_Adat_Set001_170255, QcReference_170255,
#> ℹ Cal_Example_Adat_Set002, CalQcRatio_Example_Adat_Set002_170255,
#> ℹ Dilution2
#> ── Header Data ─────────────────────────────────────────────────────────────────
#> # A tibble: 35 × 2
#>    Key                  Value                                                   
#>    <chr>                <chr>                                                   
#>  1 AdatId               GID-1234-56-789-abcdef                                  
#>  2 Version              1.2                                                     
#>  3 AssayType            PharmaServices                                          
#>  4 AssayVersion         V4                                                      
#>  5 AssayRobot           Fluent 1 L-307                                          
#>  6 Legal                Experiment details and data have been processed to prot…
#>  7 CreatedBy            PharmaServices                                          
#>  8 CreatedDate          2020-07-24                                              
#>  9 EnteredBy            Technician1                                             
#> 10 ExpDate              2020-06-18, 2020-07-20                                  
#> 11 GeneratedBy          Px (Build:  : ), Canopy_0.1.1                           
#> 12 RunNotes             2 columns ('Age' and 'Sex') have been added to this ADA…
#> 13 ProcessSteps         Raw RFU, Hyb Normalization, medNormInt (SampleId), plat…
#> 14 ProteinEffectiveDate 2019-08-06                                              
#> 15 StudyMatrix          EDTA Plasma                                             
#> # … with 20 more rows
#> ════════════════════════════════════════════════════════════════════════════════

class(example_data)
#> [1] "soma_adat"  "data.frame"

# Features/Analytes
head(ex_analytes, 20L)
#>  [1] "seq.10000.28"  "seq.10001.7"   "seq.10003.15"  "seq.10006.25" 
#>  [5] "seq.10008.43"  "seq.10011.65"  "seq.10012.5"   "seq.10013.34" 
#>  [9] "seq.10014.31"  "seq.10015.119" "seq.10021.1"   "seq.10022.207"
#> [13] "seq.10023.32"  "seq.10024.44"  "seq.10030.8"   "seq.10034.16" 
#> [17] "seq.10035.6"   "seq.10036.201" "seq.10037.98"  "seq.10040.63" 

# Feature info table (annotations)
ex_anno_tbl
#> # A tibble: 5,284 × 22
#>    AptName   SeqId SeqId…¹ SomaId Targe…² Target UniProt Entre…³ Entre…⁴ Organ…⁵
#>    <chr>     <chr>   <dbl> <chr>  <chr>   <chr>  <chr>   <chr>   <chr>   <chr>  
#>  1 seq.1000… 1000…       3 SL019… Beta-c… CRBB2  P43320  "1415"  "CRYBB… Human  
#>  2 seq.1000… 1000…       3 SL002… RAF pr… c-Raf  P04049  "5894"  "RAF1"  Human  
#>  3 seq.1000… 1000…       3 SL019… Zinc f… ZNF41  P51814  "7592"  "ZNF41" Human  
#>  4 seq.1000… 1000…       3 SL019… ETS do… ELK1   P19419  "2002"  "ELK1"  Human  
#>  5 seq.1000… 1000…       3 SL019… Guanyl… GUC1A  P43080  "2978"  "GUCA1… Human  
#>  6 seq.1001… 1001…       3 SL019… Inosit… OCRL   Q01968  "4952"  "OCRL"  Human  
#>  7 seq.1001… 1001…       3 SL014… SAM po… SPDEF  O95238  "25803" "SPDEF" Human  
#>  8 seq.1001… 1001…       3 SL025… Fc_MOU… Fc_MO… Q99LC4  ""      ""      Mouse  
#>  9 seq.1001… 1001…       3 SL007… Zinc f… SLUG   O43623  "6591"  "SNAI2" Human  
#> 10 seq.1001… 1001…       3 SL014… Voltag… KCAB2  Q13303  "8514"  "KCNAB… Human  
#> # … with 5,274 more rows, 12 more variables: Units <chr>, Type <chr>,
#> #   Dilution <chr>, PlateScale_Reference <dbl>, CalReference <dbl>,
#> #   Cal_Example_Adat_Set001 <dbl>, ColCheck <chr>,
#> #   CalQcRatio_Example_Adat_Set001_170255 <dbl>, QcReference_170255 <dbl>,
#> #   Cal_Example_Adat_Set002 <dbl>, CalQcRatio_Example_Adat_Set002_170255 <dbl>,
#> #   Dilution2 <dbl>, and abbreviated variable names ¹​SeqIdVersion,
#> #   ²​TargetFullName, ³​EntrezGeneID, ⁴​EntrezGeneSymbol, ⁵​Organism

# Search via `filter()`
dplyr::filter(ex_anno_tbl, grepl("^MMP", Target))
#> # A tibble: 15 × 22
#>    AptName   SeqId SeqId…¹ SomaId Targe…² Target UniProt Entre…³ Entre…⁴ Organ…⁵
#>    <chr>     <chr>   <dbl> <chr>  <chr>   <chr>  <chr>   <chr>   <chr>   <chr>  
#>  1 seq.1541… 1541…       3 SL012… Matrix… MMP20  O60882  9313    MMP20   Human  
#>  2 seq.2579… 2579…       5 SL000… Matrix… MMP-9  P14780  4318    MMP9    Human  
#>  3 seq.2788… 2788…       1 SL000… Strome… MMP-3  P08254  4314    MMP3    Human  
#>  4 seq.2789… 2789…       2 SL000… Matril… MMP-7  P09237  4316    MMP7    Human  
#>  5 seq.2838… 2838…       1 SL003… Matrix… MMP-17 Q9ULZ9  4326    MMP17   Human  
#>  6 seq.4160… 4160…       1 SL000… 72 kDa… MMP-2  P08253  4313    MMP2    Human  
#>  7 seq.4496… 4496…       2 SL000… Macrop… MMP-12 P39900  4321    MMP12   Human  
#>  8 seq.4924… 4924…       1 SL000… Inters… MMP-1  P03956  4312    MMP1    Human  
#>  9 seq.4925… 4925…       2 SL000… Collag… MMP-13 P45452  4322    MMP13   Human  
#> 10 seq.5002… 5002…       1 SL002… Matrix… MMP-14 P50281  4323    MMP14   Human  
#> 11 seq.5268… 5268…       3 SL003… Matrix… MMP-16 P51512  4325    MMP16   Human  
#> 12 seq.6425… 6425…       3 SL007… Matrix… MMP19  Q99542  4327    MMP19   Human  
#> 13 seq.8479… 8479…       3 SL000… Strome… MMP-10 P09238  4319    MMP10   Human  
#> 14 seq.9172… 9172…       3 SL000… Neutro… MMP-8  P22894  4317    MMP8    Human  
#> 15 seq.9719… 9719…       3 SL003… Matrix… MMP-16 P51512  4325    MMP16   Human  
#> # … with 12 more variables: Units <chr>, Type <chr>, Dilution <chr>,
#> #   PlateScale_Reference <dbl>, CalReference <dbl>,
#> #   Cal_Example_Adat_Set001 <dbl>, ColCheck <chr>,
#> #   CalQcRatio_Example_Adat_Set001_170255 <dbl>, QcReference_170255 <dbl>,
#> #   Cal_Example_Adat_Set002 <dbl>, CalQcRatio_Example_Adat_Set002_170255 <dbl>,
#> #   Dilution2 <dbl>, and abbreviated variable names ¹​SeqIdVersion,
#> #   ²​TargetFullName, ³​EntrezGeneID, ⁴​EntrezGeneSymbol, ⁵​Organism

# Lookup table -> targets
# MMP-9
ex_target_names$seq.2579.17
#> [1] "Matrix metalloproteinase-9"

# gender hormone FSH
tapply(example_data$seq.3032.11, example_data$Sex, median)
#>      F      M 
#> 3358.1  556.4 

# gender hormone LH
tapply(example_data$seq.2953.31, example_data$Sex, median)
#>      F      M 
#> 2693.8  883.6 

# Target lookup
ex_target_names$seq.2953.31     # tab-completion at console
#> [1] "Luteinizing hormone"

# Sample Type/Source
table(example_data$SampleType)
#> 
#>     Buffer Calibrator         QC     Sample 
#>          6         10          6        170 

# Sex/Gender Variable
table(example_data$Sex)
#> 
#>  F  M 
#> 85 85 

# Age Variable
summary(example_data$Age)
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max.    NA's 
#>   18.00   46.00   55.00   55.66   67.00   77.00      22