Get Analytes

Return the feature names (i.e. the column names for SOMAmer reagent analytes) from a soma_adat, data frame, matrix, list, or character vector. S3 methods exist for these classes.

getAnalytes(x, n = FALSE, rm.controls = FALSE)

getMeta(x, n = FALSE)

getFeatures(x, n = FALSE, rm.controls = FALSE)

Arguments

x
Typically a soma_adat class object created using read_adat(). Alternatively, a character vector with elements containing SeqIds corresponding to analyte features.

n
Logical. Return an integer corresponding to the length of the features?

rm.controls
Logical. Should all control and non-human analytes (e.g. HybControls, Non-Human, Non-Biotin, Spuriomer) be removed from the returned value?

Value

A character vector of ADAT feature ("analyte") names. If n = TRUE, an integer corresponding to the length of the feature names.

Either a character vector of ADAT meta data names or an integer number the length of the meta data names (if n = TRUE).

Functions

getMeta: Return a character vector of string names of non-analyte feature columns/variables.

getFeatures: renamed in SomaDataIO v5.1.0. Exported (with life-cycle warning) to maintain backward compatibility. Please adjust your code accordingly.

See also

Author

Stu Field

Examples

apts <- getAnalytes(example_data)
head(apts)
#> [1] "seq.10000.28" "seq.10001.7"  "seq.10003.15" "seq.10006.25" "seq.10008.43"
#> [6] "seq.10011.65"
getAnalytes(example_data, n = TRUE)
#> [1] 5284

# vector string
bb <- getAnalytes(names(example_data))
all.equal(apts, bb)
#> [1] TRUE

# create some control sequences
#           Spuriomer      HybControl
apts2 <- c("seq.2053.2", "seq.2171.12", head(apts))
apts2
#> [1] "seq.2053.2"   "seq.2171.12"  "seq.10000.28" "seq.10001.7"  "seq.10003.15"
#> [6] "seq.10006.25" "seq.10008.43" "seq.10011.65"
no_crtl <- getAnalytes(apts2, rm.controls = TRUE)
no_crtl
#> [1] "seq.10000.28" "seq.10001.7"  "seq.10003.15" "seq.10006.25" "seq.10008.43"
#> [6] "seq.10011.65"
setdiff(apts2, no_crtl)
#> [1] "seq.2053.2"  "seq.2171.12"

# getMeta()
mvec <- getMeta(example_data)
head(mvec, 10)
#>  [1] "PlateId"         "PlateRunDate"    "ScannerID"       "PlatePosition"  
#>  [5] "SlideId"         "Subarray"        "SampleId"        "SampleType"     
#>  [9] "PercentDilution" "SampleMatrix"   
getMeta(example_data, n = TRUE)
#> [1] 34

# test data.frame and character S3 methods
identical(getMeta(example_data), getMeta(names(example_data))) # TRUE
#> [1] TRUE