Group Generics for soma_adat Class Objects

S3 group generic methods to apply group specific prototype functions to the RFU data only of soma_adat objects. The clinical meta data are not transformed and remain unmodified in the returned object (Math() and Ops()) or are ignored for the Summary() group. See groupGeneric().

# S3 method for soma_adat
Math(x, ...)

antilog(x, base = 10)

# S3 method for soma_adat
Ops(e1, e2 = NULL)

# S3 method for soma_adat
==(e1, e2)

# S3 method for soma_adat
Summary(..., na.rm = FALSE)

Arguments

x

The soma_adat class object to perform the transformation.

...

Additional arguments passed to the various group generics as appropriate.

base

A positive or complex number: the base with respect to which logarithms are computed.

e1, e2

Objects.

na.rm

Logical. Should missing values be removed?

Value

A soma_adat object with the same dimensions of the input object with the feature columns transformed by the specified generic.

Functions

antilog(): Performs the inverse or anti-log transform for a numeric vector of soma_adat object. note: default is base = 10, which differs from the log() default base e.

Ops(soma_adat): Performs binary mathematical operations on class soma_adat. See Ops().

== : Compares left- and right-hand sides of the operator unless the RHS is also a soma_adat, in which case diffAdats() is invoked.

Summary(soma_adat): Performs summary calculations on class soma_adat. See Summary().

Math

Group members:

#>  [1] "abs"      "acos"     "acosh"    "asin"     "asinh"    "atan"    
#>  [7] "atanh"    "ceiling"  "cos"      "cosh"     "cospi"    "cummax"  
#> [13] "cummin"   "cumprod"  "cumsum"   "digamma"  "exp"      "expm1"   
#> [19] "floor"    "gamma"    "lgamma"   "log"      "log10"    "log1p"   
#> [25] "log2"     "sign"     "sin"      "sinh"     "sinpi"    "sqrt"    
#> [31] "tan"      "tanh"     "tanpi"    "trigamma" "trunc"

Commonly used generics of this group include:

log(), log10(), log2(), antilog(), abs(), sign(), floor(), sqrt(), exp()

Ops

Group members:

#>  [1] "+"   "-"   "*"   "^"   "%%"  "%/%" "/"   "=="  ">"   "<"   "!="  "<=" 
#> [13] ">="

Note that for the `==` method if the RHS is also a soma_adat, diffAdats() is invoked which compares LHS vs. RHS. Commonly used generics of this group include:

+, -, *, /, ^, ==, >, <

See also

Author

Stu Field

Examples

example_data$seq.3343.1
#>   [1]  5282.8  2014.8  2046.6  6008.1  2198.3  1736.4  2365.1  1439.2  1923.2
#>  [10]  3027.9  2077.1  1993.5    47.5  2736.3  2532.1  3303.4  2541.4  6327.7
#>  [19]  4006.8  2509.7  3643.7  3205.4  4353.5  3603.0  1881.8  4022.0  4326.9
#>  [28]   858.0   943.2   967.6  2372.6  1271.6  2433.1  1805.9  3064.3  3034.9
#>  [37]  4027.5  3094.5  2789.3  2849.8 15314.0  1292.2    47.2  2549.8  2988.0
#>  [46]  3273.5  3380.2  1999.5  2653.5  3209.2  3580.2  4060.1  1939.8  3697.1
#>  [55]  1436.6  2119.3  4111.3  1230.9  2648.5  2675.9  2241.4  2613.1  4811.2
#>  [64]  1417.7  1571.2  3569.7  2714.9  3098.5  3083.1  3700.5  2176.4  3364.7
#>  [73]  4787.1  3459.4  2997.3  2621.2  3149.5  3027.5  3421.2  2259.5  2500.2
#>  [82]  9027.4  2006.0  5660.0  9066.4  1287.7    45.8  1245.8  2109.7  3438.1
#>  [91]  3826.7  3448.6  2077.6  1553.9  2951.0  2180.8  1838.9  3018.7  2623.3
#> [100]  5663.1  1828.3  3493.7  2278.1  1527.5  1824.3  5389.4  3024.9  2278.8
#> [109]  3041.3  6150.7  3317.7 13048.4  2733.9  3191.8  3865.9  2710.9  3815.3
#> [118]  2564.4  3991.1  1747.3  9328.0  4955.1  4224.8  2281.7  1887.9  4500.0
#> [127]  1347.4  1737.5  3008.0  3330.9  2251.3  3785.3  2585.4  2016.0  1173.4
#> [136]  2569.8  1888.7  4750.2  6805.1  3043.1  3475.9  3804.9    38.7  1891.8
#> [145]  2225.9  4467.9  1929.1  3616.1  1911.3  3253.3    38.9  2448.3  3140.5
#> [154]  2627.3  2746.9  1561.5  5631.2  1540.5  3317.7  2182.3  2131.6  1709.4
#> [163]  1839.8  5028.3  5104.5  4564.6  2815.2  2727.6  2000.5  4259.4  2899.4
#> [172]  1874.8    43.4  2034.7  6400.8  1855.7  3850.4  1803.2  1182.4  3034.9
#> [181]  2962.1  1425.9  2598.0  5532.3  3482.9  6883.1  1026.4  2126.6  1388.7
#> [190]  2084.4  1358.0  1867.5

# Math Generics:
# -------------
# log-transformation
a <- log(example_data)
a$seq.3343.1
#>   [1] 8.572212 7.608275 7.623935 8.700864 7.695440 7.459569 7.768576 7.271843
#>   [9] 7.561746 8.015625 7.638728 7.597647 3.860730 7.914362 7.836804 8.102708
#>  [17] 7.840470 8.752692 8.295748 7.827919 8.200755 8.072592 8.378735 8.189522
#>  [25] 7.539984 8.299535 8.372607 6.754604 6.849278 6.874819 7.771742 7.148031
#>  [33] 7.796921 7.498814 8.027574 8.017934 8.300901 8.037382 7.933546 7.955004
#>  [41] 9.636523 7.164101 3.854394 7.843770 8.002360 8.093615 8.125690 7.600652
#>  [49] 7.883635 8.073777 8.183174 8.308963 7.570340 8.215304 7.270034 7.658841
#>  [57] 8.321495 7.115501 7.881749 7.892041 7.714856 7.868293 8.478702 7.256791
#>  [65] 7.359595 8.180237 7.906510 8.038673 8.033691 8.216223 7.685427 8.121094
#>  [73] 8.473680 8.148850 8.005467 7.871388 8.054999 8.015492 8.137747 7.722899
#>  [81] 7.824126 9.108020 7.603898 8.641179 9.112331 7.160613 3.824284 7.127533
#>  [89] 7.654301 8.142674 8.249758 8.145724 7.638969 7.348523 7.989899 7.687447
#>  [97] 7.516923 8.012582 7.872188 8.641727 7.511142 8.158717 7.731097 7.331388
#> [105] 7.508952 8.592189 8.014633 7.731404 8.020040 8.724321 8.107027 9.476421
#> [113] 7.913484 8.068340 8.259950 7.905036 8.246775 7.849480 8.291822 7.465827
#> [121] 9.140776 8.508173 8.348727 7.732676 7.543220 8.411833 7.205932 7.460203
#> [129] 8.009031 8.110998 7.719263 8.238880 7.857636 7.608871 7.067661 7.851583
#> [137] 7.543644 8.465942 8.825428 8.020632 8.153609 8.244045 3.655840 7.545284
#> [145] 7.707917 8.404674 7.564809 8.193151 7.555539 8.087425 3.660994 7.803149
#> [153] 8.052137 7.873712 7.918228 7.353402 8.636078 7.339862 8.107027 7.688135
#> [161] 7.664628 7.443898 7.517412 8.522837 8.537878 8.426086 7.942789 7.911177
#> [169] 7.601152 8.356884 7.972259 7.536257 3.770459 7.618104 8.764178 7.526017
#> [177] 8.255932 7.497318 7.075302 8.017934 7.993654 7.262558 7.862497 8.618359
#> [185] 8.155621 8.836824 6.933813 7.662280 7.236123 7.642236 7.213768 7.532356

b <- log10(example_data)
b$seq.3343.1
#>   [1] 3.722864 3.304232 3.311033 3.778737 3.342087 3.239650 3.373850 3.158121
#>   [9] 3.284024 3.481142 3.317457 3.299616 1.676694 3.437164 3.403481 3.518961
#>  [17] 3.405073 3.801246 3.602798 3.399622 3.561543 3.505882 3.638839 3.556664
#>  [25] 3.274573 3.604442 3.636177 2.933487 2.974604 2.985696 3.375225 3.104351
#>  [33] 3.386160 3.256694 3.486331 3.482144 3.605036 3.490590 3.445495 3.454814
#>  [41] 4.185089 3.111330 1.673942 3.406506 3.475381 3.515012 3.528942 3.300921
#>  [49] 3.423819 3.506397 3.553907 3.608537 3.287757 3.567861 3.157336 3.326192
#>  [57] 3.613979 3.090223 3.423000 3.427470 3.350519 3.417156 3.682253 3.151584
#>  [65] 3.196231 3.552632 3.433754 3.491152 3.488988 3.568260 3.337739 3.526946
#>  [73] 3.680072 3.539001 3.476730 3.418500 3.498242 3.481084 3.534178 3.354012
#>  [81] 3.397975 3.955563 3.302331 3.752816 3.957435 3.109815 1.660865 3.095448
#>  [89] 3.324221 3.536319 3.582824 3.537643 3.317562 3.191423 3.469969 3.338616
#>  [97] 3.264558 3.479820 3.418848 3.753054 3.262047 3.543286 3.357573 3.183981
#> [105] 3.261096 3.731540 3.480711 3.357706 3.483059 3.788925 3.520837 4.115557
#> [113] 3.436783 3.504036 3.587251 3.433113 3.581529 3.408986 3.601093 3.242367
#> [121] 3.969789 3.695052 3.625806 3.358259 3.275979 3.653213 3.129497 3.239925
#> [129] 3.478278 3.522562 3.352433 3.578100 3.412528 3.304491 3.069446 3.409899
#> [137] 3.276163 3.676712 3.832835 3.483316 3.541067 3.580343 1.587711 3.276875
#> [145] 3.347506 3.650103 3.285355 3.558240 3.281329 3.512324 1.589950 3.388865
#> [153] 3.496999 3.419510 3.438843 3.193542 3.750601 3.187662 3.520837 3.338914
#> [161] 3.328706 3.232844 3.264771 3.701421 3.707953 3.659403 3.449509 3.435781
#> [169] 3.301139 3.629348 3.462308 3.272955 1.637490 3.308500 3.806234 3.268508
#> [177] 3.585506 3.256044 3.072764 3.482144 3.471600 3.154089 3.414639 3.742906
#> [185] 3.541941 3.837784 3.011317 3.327686 3.142608 3.318981 3.132900 3.271261
isTRUE(all.equal(b, log(example_data, base = 10)))
#> [1] TRUE

# floor
c <- floor(example_data)
c$seq.3343.1
#>   [1]  5282  2014  2046  6008  2198  1736  2365  1439  1923  3027  2077  1993
#>  [13]    47  2736  2532  3303  2541  6327  4006  2509  3643  3205  4353  3603
#>  [25]  1881  4022  4326   858   943   967  2372  1271  2433  1805  3064  3034
#>  [37]  4027  3094  2789  2849 15314  1292    47  2549  2988  3273  3380  1999
#>  [49]  2653  3209  3580  4060  1939  3697  1436  2119  4111  1230  2648  2675
#>  [61]  2241  2613  4811  1417  1571  3569  2714  3098  3083  3700  2176  3364
#>  [73]  4787  3459  2997  2621  3149  3027  3421  2259  2500  9027  2006  5660
#>  [85]  9066  1287    45  1245  2109  3438  3826  3448  2077  1553  2951  2180
#>  [97]  1838  3018  2623  5663  1828  3493  2278  1527  1824  5389  3024  2278
#> [109]  3041  6150  3317 13048  2733  3191  3865  2710  3815  2564  3991  1747
#> [121]  9328  4955  4224  2281  1887  4500  1347  1737  3008  3330  2251  3785
#> [133]  2585  2016  1173  2569  1888  4750  6805  3043  3475  3804    38  1891
#> [145]  2225  4467  1929  3616  1911  3253    38  2448  3140  2627  2746  1561
#> [157]  5631  1540  3317  2182  2131  1709  1839  5028  5104  4564  2815  2727
#> [169]  2000  4259  2899  1874    43  2034  6400  1855  3850  1803  1182  3034
#> [181]  2962  1425  2598  5532  3482  6883  1026  2126  1388  2084  1358  1867

# square-root
d <- sqrt(example_data)
d$seq.3343.1
#>   [1]  72.682873  44.886524  45.239363  77.511935  46.886032  41.670133
#>   [7]  48.632294  37.936790  43.854304  55.026357  45.575213  44.648628
#>  [13]   6.892024  52.309655  50.319976  57.475212  50.412300  79.546842
#>  [19]  63.299289  50.096906  60.363068  56.616252  65.981058  60.024995
#>  [25]  43.379719  63.419240  65.779176  29.291637  30.711561  31.106269
#>  [31]  48.709342  35.659501  49.326463  42.495882  55.356120  55.089926
#>  [37]  63.462587  55.628230  52.813824  53.383518 123.749747  35.947183
#>  [43]   6.870226  50.495544  54.662601  57.214509  58.139487  44.715769
#>  [49]  51.512134  56.649801  59.834772  63.718914  44.043161  60.803783
#>  [55]  37.902507  46.035856  64.119420  35.084184  51.463579  51.729102
#>  [61]  47.343426  51.118490  69.362814  37.652357  39.638365  59.746966
#>  [67]  52.104702  55.664172  55.525670  60.831735  46.651902  58.006034
#>  [73]  69.188872  58.816664  54.747603  51.197656  56.120406  55.022723
#>  [79]  58.491025  47.534198  50.002000  95.012631  44.788391  75.232971
#>  [85]  95.217645  35.884537   6.767570  35.295892  45.931471  58.635314
#>  [91]  61.860327  58.724782  45.580698  39.419538  54.323107  46.699036
#>  [97]  42.882397  54.942697  51.218161  75.253571  42.758625  59.107529
#> [103]  47.729446  39.083244  42.711825  73.412533  54.999091  47.736778
#> [109]  55.147983  78.426399  57.599479 114.229593  52.286710  56.496018
#> [115]  62.176362  52.066304  61.768115  50.639905  63.175153  41.800718
#> [121]  96.581572  70.392471  64.998462  47.767144  43.449971  67.082039
#> [127]  36.706948  41.683330  54.845237  57.713950  47.447866  61.524792
#> [133]  50.846829  44.899889  34.254927  50.693195  43.459176  68.921695
#> [139]  82.493030  55.164300  58.956764  61.683871   6.220932  43.494827
#> [145]  47.179445  66.842352  43.921521  60.134017  43.718417  57.037707
#> [151]   6.236986  49.480299  56.040164  51.257195  52.410877  39.515820
#> [157]  75.041322  39.249204  57.599479  46.715094  46.169254  41.344891
#> [163]  42.892890  70.910507  71.445784  67.561824  53.058458  52.226430
#> [169]  44.726949  65.264079  53.846077  43.298961   6.587868  45.107649
#> [175]  80.005000  43.077837  62.051591  42.464102  34.386044  55.089926
#> [181]  54.425178  37.761091  50.970580  74.379433  59.016099  82.964450
#> [187]  32.037478  46.115073  37.265265  45.655230  36.851052  43.214581

# rounding
e <- round(example_data)
e$seq.3343.1
#>   [1]  5283  2015  2047  6008  2198  1736  2365  1439  1923  3028  2077  1994
#>  [13]    48  2736  2532  3303  2541  6328  4007  2510  3644  3205  4354  3603
#>  [25]  1882  4022  4327   858   943   968  2373  1272  2433  1806  3064  3035
#>  [37]  4028  3094  2789  2850 15314  1292    47  2550  2988  3274  3380  2000
#>  [49]  2654  3209  3580  4060  1940  3697  1437  2119  4111  1231  2648  2676
#>  [61]  2241  2613  4811  1418  1571  3570  2715  3098  3083  3700  2176  3365
#>  [73]  4787  3459  2997  2621  3150  3028  3421  2260  2500  9027  2006  5660
#>  [85]  9066  1288    46  1246  2110  3438  3827  3449  2078  1554  2951  2181
#>  [97]  1839  3019  2623  5663  1828  3494  2278  1528  1824  5389  3025  2279
#> [109]  3041  6151  3318 13048  2734  3192  3866  2711  3815  2564  3991  1747
#> [121]  9328  4955  4225  2282  1888  4500  1347  1738  3008  3331  2251  3785
#> [133]  2585  2016  1173  2570  1889  4750  6805  3043  3476  3805    39  1892
#> [145]  2226  4468  1929  3616  1911  3253    39  2448  3140  2627  2747  1562
#> [157]  5631  1540  3318  2182  2132  1709  1840  5028  5104  4565  2815  2728
#> [169]  2000  4259  2899  1875    43  2035  6401  1856  3850  1803  1182  3035
#> [181]  2962  1426  2598  5532  3483  6883  1026  2127  1389  2084  1358  1868

# inverse log
antilog(1:4)
#> [1]    10   100  1000 10000

alog <- antilog(b)
all.equal(example_data, alog)    # return `b` -> linear space
#> [1] TRUE

# Ops Generics:
# -------------
plus1 <- example_data + 1
times2 <- example_data * 2

sq <- example_data^2
all.equal(sqrt(sq), example_data)
#> [1] TRUE

gt100k <- example_data > 100000
gt100k
#> ══ SomaScan Data ═══════════════════════════════════════════════════════════════
#>      Attributes intact    ✓
#>      Rows                 192
#>      Columns              5318
#>      Clinical Data        34
#>      Features             5284
#> ── Column Meta ─────────────────────────────────────────────────────────────────
#> ℹ SeqId, SeqIdVersion, SomaId, TargetFullName, Target, UniProt,
#> ℹ EntrezGeneID, EntrezGeneSymbol, Organism, Units, Type, Dilution,
#> ℹ PlateScale_Reference, CalReference, Cal_Example_Adat_Set001,
#> ℹ ColCheck, CalQcRatio_Example_Adat_Set001_170255, QcReference_170255,
#> ℹ Cal_Example_Adat_Set002, CalQcRatio_Example_Adat_Set002_170255,
#> ℹ Dilution2
#> ── Tibble ──────────────────────────────────────────────────────────────────────
#> # A tibble: 192 × 5,319
#>    row_names     PlateId Plate…¹ Scann…² Plate…³ SlideId Subar…⁴ Sampl…⁵ Sampl…⁶
#>    <chr>         <chr>   <chr>   <chr>   <chr>     <dbl>   <dbl> <chr>   <chr>  
#>  1 258495800012… Exampl… 2020-0… SG1521… H9      2.58e11       3 1       Sample 
#>  2 258495800004… Exampl… 2020-0… SG1521… H8      2.58e11       7 2       Sample 
#>  3 258495800010… Exampl… 2020-0… SG1521… H7      2.58e11       8 3       Sample 
#>  4 258495800003… Exampl… 2020-0… SG1521… H6      2.58e11       4 4       Sample 
#>  5 258495800009… Exampl… 2020-0… SG1521… H5      2.58e11       4 5       Sample 
#>  6 258495800012… Exampl… 2020-0… SG1521… H4      2.58e11       8 6       Sample 
#>  7 258495800001… Exampl… 2020-0… SG1521… H3      2.58e11       3 7       Sample 
#>  8 258495800004… Exampl… 2020-0… SG1521… H2      2.58e11       8 8       Sample 
#>  9 258495800001… Exampl… 2020-0… SG1521… H12     2.58e11       8 9       Sample 
#> 10 258495800004… Exampl… 2020-0… SG1521… H11     2.58e11       3 170261  Calibr…
#> # … with 182 more rows, 5,310 more variables: PercentDilution <int>,
#> #   SampleMatrix <chr>, Barcode <lgl>, Barcode2d <chr>, SampleName <lgl>,
#> #   SampleNotes <lgl>, AliquotingNotes <lgl>, SampleDescription <chr>,
#> #   AssayNotes <lgl>, TimePoint <lgl>, …, and abbreviated variable names
#> #   ¹​PlateRunDate, ²​ScannerID, ³​PlatePosition, ⁴​Subarray, ⁵​SampleId, …
#> ════════════════════════════════════════════════════════════════════════════════

example_data == example_data   # invokes diffAdats()
#> ══ Checking ADAT attributes & characteristics ══════════════════════════════════
#> • Attribute names are identical       ✓
#> • Attributes are identical            ✓
#> • ADAT dimensions are identical       ✓
#> • ADAT row names are identical        ✓
#> • ADATs contain identical Features    ✓
#> • ADATs contain same Meta Fields      ✓
#> ── Checking the data matrix ────────────────────────────────────────────────────
#> • All Clinical data is identical      ✓
#> • All Feature data is identical       ✓
#> ════════════════════════════════════════════════════════════════════════════════
#> NULL

# Summary Generics:
# -------------
sum(example_data)
#> [1] 5031475869

any(example_data < 100)  # low RFU analytes
#> [1] TRUE

sum(example_data < 100)  # how many
#> [1] 41721

min(example_data)
#> [1] 1.2

min(example_data, 0)
#> [1] 0

max(example_data)
#> [1] 264534.5

max(example_data, 1e+7)
#> [1] 1e+07

range(example_data)
#> [1]      1.2 264534.5