Lift an ADAT Between Assay Versions

The SomaScan platform continually improves its technical processes between assay versions; from changing reagents, liquid handling equipment, and increased analyte content. However, these upgrades can result in minute differences in RFU space for a given analyte, requiring a calibration (aka "lifting" or "bridging") to bring RFUs into a comparable space. This is accomplished by applying an analyte-specific scalar to each analyte RFU (ADAT column). The scalar values themselves are typically provided via *.xlsx file, which can be parsed via read_annotations(). See Details.

Usage

lift_adat(adat, anno.tbl)

Arguments

adat

A soma_adat class object.

anno.tbl

A table of annotations, typically the result of a call to read_annotations().

Value

A "lifted" soma_adat object corresponding to the scaling reference in the anno.tbl. RFU values are rounded to 1 decimal to match standard SomaScan delivery format.

Details

Lifting between various versions requires a specific annotations file containing scalars specific to desired lifting direction. For example, to "lift" between v4.1 -> v4.0, you must be working with SomaScan data in v4.1 space and an annotations file containing scalars to convert to v4.0. Likewise, "lifting" from v4.0 -> v4.1 requires a separate annotations file and a soma_adat from SomaScan v4.0.

Examples

# `example_data` is SomaScan V4
adat <- head(example_data, 3L)

# read in version specific annotations file
# containing scaling values between assay versions
if (FALSE) {
tbl <- read_annotations("path/to/annotations_file.xlsx")
}

# mock annotations table in lieu of `*.xlsx` file
tbl <- tibble::tibble(SeqId = getSeqId(getAnalytes(adat)),
                     "Plasma Scalar v4.0 to v4.1" = 1)   # scale by 1.0
# usually performed inside `read_annotations()`
# assign valid testing version to annotations table
attr(tbl, "version") <- "SL-99999999-rev99-1999-01"

# perform the 'lifting'
lifted <- lift_adat(adat, tbl)

# `tbl` contained all scalars = 1.0 (same RFUs)
all.equal(adat, lifted, check.attributes = FALSE)
#> [1] TRUE

# attributes updated to reflect the 'lift'
attr(lifted, "Header")$HEADER$ProcessSteps
#> [1] "Raw RFU, Hyb Normalization, medNormInt (SampleId), plateScale, Calibration, anmlQC, qcCheck, anmlSMP, Annotation Lift (v4 to v4.1)"
attr(lifted, "Header")$HEADER$SignalSpace
#> [1] "v4.1"