Read (Load) SomaLogic ADATs

The parse and load a *.adat file as a data.frame-like object into an R workspace environment. The class of the returned object is a soma_adat object.

read.adat() is a convenient backward compatibility alias for read_adat() to enable use of older versions of SomaDataIO. It will likely never go away completely, but you strongly encouraged to shift your code to use read_adat().

is.soma_adat() checks whether an object is of class soma_adat. See inherits().

read_adat(file, debug = FALSE, verbose = getOption("verbose"), ...)

read.adat(file, debug = FALSE, verbose = getOption("verbose"), ...)

is.soma_adat(x)

Arguments

file

Character. The elaborated path and file name of the *.adat file to be loaded into an R workspace.

debug

Logical. Used for debugging and development of an ADAT that fails to load, particularly out-of-spec, poorly modified, or legacy ADATs.

verbose

Logical. Should the function call be run in verbose mode, printing relevant diagnostic call information to the console.

...

Additional arguments passed ultimately to read.delim(), or additional arguments passed to either other S3 print or summary methods as required by those generics.

x

An R object to test.

Value

A data.frame-like object of class soma_adat

consisting of SomaLogic RFU (feature) data and clinical meta data as columns, and samples as rows. Row names are labeled with the unique ID "SlideId_Subarray" concatenation. The sections of the ADAT header (e.g., "Header.Meta", "Col.Meta", ...) are stored as attributes (e.g. attributes(x)$Header.Meta).

Logical. Whether x inherits from class soma_adat.

See also

Author

Stu Field

Examples

f <- system.file("extdata", "example_data10.adat",
                 package = "SomaDataIO", mustWork = TRUE)
my_adat <- read_adat(f)

is.soma_adat(my_adat)
#> [1] TRUE