Read (Load) SomaLogic ADATs

The parse and load a *.adat file as a data.frame-like object into an R workspace environment. The class of the returned object is a soma_adat object.

read.adat() is a convenient alias for read_adat() designed to enable backward compatibility to older versions of SomaDataIO.

is.soma_adat() checks whether an object is of class soma_adat. See inherits().

read_adat(file, debug = FALSE, verbose = getOption("verbose"), ...)

read.adat(file, debug = FALSE, verbose = getOption("verbose"), ...)

is.soma_adat(x)

Arguments

file

Character. The elaborated path and file name of the *.adat file to be loaded into an R workspace.

debug

Logical. Used for debugging and development of an ADAT that fails to load, particularly out-of-spec, poorly modified, or legacy ADATs.

verbose

Logical. Should the function call be run in verbose mode, printing relevant diagnostic call information to the console.

...

Additional arguments passed ultimately to read.delim(), or additional arguments passed to either other S3 print or summary methods as required by those generics.

x

An R object to test.

Value

A data.frame-like object of class soma_adat

consisting of SomaLogic RFU (feature) data and clinical meta data as columns, and samples as rows. Row names are labeled with the unique ID "SlideId_Subarray" concatenation. The sections of the ADAT header (e.g., "Header.Meta", "Col.Meta", ...) are stored as attributes (e.g. attributes(x)$Header.Meta).

A logical indicating whether x inherits from class soma_adat.

See also

Author

Stu Field

Examples

f <- system.file("example", "example_data.adat",
                 package = "SomaDataIO", mustWork = TRUE)
my_adat <- read_adat(f)
is.soma_adat(my_adat)
#> [1] TRUE

# S3 print method
my_adat                             # redirect uses `tibble` method
#> ══ SomaScan Data ═══════════════════════════════════════════════════════════════
#>      Attributes intact    ✓
#>      Rows                 192
#>      Columns              5318
#>      Clinical Data        34
#>      Features             5284
#> ── Column Meta ─────────────────────────────────────────────────────────────────
#> ℹ SeqId, SeqIdVersion, SomaId, TargetFullName, Target, UniProt,
#> ℹ EntrezGeneID, EntrezGeneSymbol, Organism, Units, Type, Dilution,
#> ℹ PlateScale_Reference, CalReference, Cal_Example_Adat_Set001,
#> ℹ ColCheck, CalQcRatio_Example_Adat_Set001_170255, QcReference_170255,
#> ℹ Cal_Example_Adat_Set002, CalQcRatio_Example_Adat_Set002_170255,
#> ℹ Dilution2
#> ── Tibble ──────────────────────────────────────────────────────────────────────
#> # A tibble: 192 × 5,319
#>    row_names     PlateId Plate…¹ Scann…² Plate…³ SlideId Subar…⁴ Sampl…⁵ Sampl…⁶
#>    <chr>         <chr>   <chr>   <chr>   <chr>     <dbl>   <dbl> <chr>   <chr>  
#>  1 258495800012… Exampl… 2020-0… SG1521… H9      2.58e11       3 1       Sample 
#>  2 258495800004… Exampl… 2020-0… SG1521… H8      2.58e11       7 2       Sample 
#>  3 258495800010… Exampl… 2020-0… SG1521… H7      2.58e11       8 3       Sample 
#>  4 258495800003… Exampl… 2020-0… SG1521… H6      2.58e11       4 4       Sample 
#>  5 258495800009… Exampl… 2020-0… SG1521… H5      2.58e11       4 5       Sample 
#>  6 258495800012… Exampl… 2020-0… SG1521… H4      2.58e11       8 6       Sample 
#>  7 258495800001… Exampl… 2020-0… SG1521… H3      2.58e11       3 7       Sample 
#>  8 258495800004… Exampl… 2020-0… SG1521… H2      2.58e11       8 8       Sample 
#>  9 258495800001… Exampl… 2020-0… SG1521… H12     2.58e11       8 9       Sample 
#> 10 258495800004… Exampl… 2020-0… SG1521… H11     2.58e11       3 170261  Calibr…
#> # … with 182 more rows, 5,310 more variables: PercentDilution <int>,
#> #   SampleMatrix <chr>, Barcode <lgl>, Barcode2d <chr>, SampleName <lgl>,
#> #   SampleNotes <lgl>, AliquotingNotes <lgl>, SampleDescription <chr>,
#> #   AssayNotes <lgl>, TimePoint <lgl>, …, and abbreviated variable names
#> #   ¹​PlateRunDate, ²​ScannerID, ³​PlatePosition, ⁴​Subarray, ⁵​SampleId, …
#> ════════════════════════════════════════════════════════════════════════════════
print(my_adat, show_header = TRUE)  # show the header info; no RFU data
#> ══ SomaScan Data ═══════════════════════════════════════════════════════════════
#>      Attributes intact    ✓
#>      Rows                 192
#>      Columns              5318
#>      Clinical Data        34
#>      Features             5284
#> ── Column Meta ─────────────────────────────────────────────────────────────────
#> ℹ SeqId, SeqIdVersion, SomaId, TargetFullName, Target, UniProt,
#> ℹ EntrezGeneID, EntrezGeneSymbol, Organism, Units, Type, Dilution,
#> ℹ PlateScale_Reference, CalReference, Cal_Example_Adat_Set001,
#> ℹ ColCheck, CalQcRatio_Example_Adat_Set001_170255, QcReference_170255,
#> ℹ Cal_Example_Adat_Set002, CalQcRatio_Example_Adat_Set002_170255,
#> ℹ Dilution2
#> ── Header Data ─────────────────────────────────────────────────────────────────
#> # A tibble: 35 × 2
#>    Key                  Value                                                   
#>    <chr>                <chr>                                                   
#>  1 AdatId               GID-1234-56-789-abcdef                                  
#>  2 Version              1.2                                                     
#>  3 AssayType            PharmaServices                                          
#>  4 AssayVersion         V4                                                      
#>  5 AssayRobot           Fluent 1 L-307                                          
#>  6 Legal                Experiment details and data have been processed to prot…
#>  7 CreatedBy            PharmaServices                                          
#>  8 CreatedDate          2020-07-24                                              
#>  9 EnteredBy            Technician1                                             
#> 10 ExpDate              2020-06-18, 2020-07-20                                  
#> 11 GeneratedBy          Px (Build:  : ), Canopy_0.1.1                           
#> 12 RunNotes             2 columns ('Age' and 'Sex') have been added to this ADA…
#> 13 ProcessSteps         Raw RFU, Hyb Normalization, medNormInt (SampleId), plat…
#> 14 ProteinEffectiveDate 2019-08-06                                              
#> 15 StudyMatrix          EDTA Plasma                                             
#> # … with 20 more rows
#> ════════════════════════════════════════════════════════════════════════════════

# write ADAT to file
fout <- tempfile(fileext = ".adat")
write_adat(my_adat, file = fout)
#> ✔ ADAT passed all checks and traps.
#> ✔ ADAT written to: '/tmp/Rtmpw7MWXH/file41d76cdcea95.adat'

# read same file back in as check
read_adat(fout)
#> ══ SomaScan Data ═══════════════════════════════════════════════════════════════
#>      Attributes intact    ✓
#>      Rows                 192
#>      Columns              5318
#>      Clinical Data        34
#>      Features             5284
#> ── Column Meta ─────────────────────────────────────────────────────────────────
#> ℹ SeqId, SeqIdVersion, SomaId, TargetFullName, Target, UniProt,
#> ℹ EntrezGeneID, EntrezGeneSymbol, Organism, Units, Type, Dilution,
#> ℹ PlateScale_Reference, CalReference, Cal_Example_Adat_Set001,
#> ℹ ColCheck, CalQcRatio_Example_Adat_Set001_170255, QcReference_170255,
#> ℹ Cal_Example_Adat_Set002, CalQcRatio_Example_Adat_Set002_170255,
#> ℹ Dilution2
#> ── Tibble ──────────────────────────────────────────────────────────────────────
#> # A tibble: 192 × 5,319
#>    row_names     PlateId Plate…¹ Scann…² Plate…³ SlideId Subar…⁴ Sampl…⁵ Sampl…⁶
#>    <chr>         <chr>   <chr>   <chr>   <chr>     <dbl>   <dbl> <chr>   <chr>  
#>  1 258495800012… Exampl… 2020-0… SG1521… H9      2.58e11       3 1       Sample 
#>  2 258495800004… Exampl… 2020-0… SG1521… H8      2.58e11       7 2       Sample 
#>  3 258495800010… Exampl… 2020-0… SG1521… H7      2.58e11       8 3       Sample 
#>  4 258495800003… Exampl… 2020-0… SG1521… H6      2.58e11       4 4       Sample 
#>  5 258495800009… Exampl… 2020-0… SG1521… H5      2.58e11       4 5       Sample 
#>  6 258495800012… Exampl… 2020-0… SG1521… H4      2.58e11       8 6       Sample 
#>  7 258495800001… Exampl… 2020-0… SG1521… H3      2.58e11       3 7       Sample 
#>  8 258495800004… Exampl… 2020-0… SG1521… H2      2.58e11       8 8       Sample 
#>  9 258495800001… Exampl… 2020-0… SG1521… H12     2.58e11       8 9       Sample 
#> 10 258495800004… Exampl… 2020-0… SG1521… H11     2.58e11       3 170261  Calibr…
#> # … with 182 more rows, 5,310 more variables: PercentDilution <int>,
#> #   SampleMatrix <chr>, Barcode <lgl>, Barcode2d <chr>, SampleName <lgl>,
#> #   SampleNotes <lgl>, AliquotingNotes <lgl>, SampleDescription <chr>,
#> #   AssayNotes <lgl>, TimePoint <lgl>, …, and abbreviated variable names
#> #   ¹​PlateRunDate, ²​ScannerID, ³​PlatePosition, ⁴​Subarray, ⁵​SampleId, …
#> ════════════════════════════════════════════════════════════════════════════════