The soma_adatClass and S3 Methods

The soma_adat data structure is the primary internal R representation of SomaScan data. A soma_adat is automatically created via read_adat() when loading a *.adat text file. It consists of a data.frame-like object with leading columns as clinical variables and SomaScan RFU data as the remaining variables. Two main attributes corresponding to analyte and SomaScan run information contained in the *.adat file are added:

Header.Meta: information about the SomaScan run, see parseHeader() or attr(x, "Header.Meta")

Col.Meta: annotations information about the SomaScan reagents/analytes, see getAnalyteInfo() or attr(x, "Col.Meta")

file_specs: parsing specifications for the ingested *.adat file

row_meta: the names of the non-RFU fields. See getMeta().

See groupGenerics() for a details on Math(), Ops(), and Summary() methods that dispatch on class soma_adat.

See reexports() for a details on re-exported S3 generics from other packages (mostly dplyr and tidyr) to enable S3 methods to be dispatched on class soma_adat.

Below is a list of all currently available S3 methods that dispatch on the soma_adat class:

##  [1] [            [[           [[<-         [<-          ==          
##  [6] $            $<-          anti_join    arrange      count       
## [11] filter       full_join    getAnalytes  getMeta      group_by    
## [16] inner_join   is_seqFormat left_join    Math         median      
## [21] mutate       Ops          print        rename       right_join  
## [26] sample_frac  sample_n     select       semi_join    separate    
## [31] slice_sample slice        summary      Summary      transform   
## [36] ungroup      unite       
## see '?methods' for accessing help and source code

S3 print() method returns summary information parsed from the object attributes, if present, followed by a dispatch to the tibble() print method. Rownames are printed as the first column in the print method only.

S3 summary() method returns the following for each column of the ADAT object containing SOMAmer data (clinical meta data is excluded):

Target (if available)

Minimum value

1st Quantile

Median

Mean

3rd Quantile

Maximum value

Standard deviation

Median absolute deviation (mad())

Interquartile range (IQR())

S3 Extract() method is used for sub-setting a soma_adat object and relies heavily on the [ method that maintains the soma_adat attributes intact and subsets the Col.Meta so that it is consistent with the newly created object.

S3 extraction via $ is fully supported, however, as opposed to the data.frame method, partial matching is not allowed for class soma_adat.

S3 extraction via [[ is fully supported, however, we restrict the usage of [[ for soma_adat. Use only a numeric index (e.g. 1L) or a character identifying the column (e.g. "SampleID"). Do not use [[i,j]] syntax with [[, use [ instead. As with $, partial matching is not allowed.

S3 assignment via [ is fully supported for class soma_adat.

S3 assignment via $ is fully supported for class soma_adat.

S3 assignment via [[ is fully supported for class soma_adat.

S3 median() is currently unavailable for the soma_adat class.

# S3 method for soma_adat
print(x, show_header = FALSE, ...)

# S3 method for soma_adat
summary(object, tbl = NULL, digits = max(3L, getOption("digits") - 3L), ...)

# S3 method for soma_adat
[(x, i, j, drop = TRUE)

# S3 method for soma_adat
$(x, name)

# S3 method for soma_adat
[[(x, i, j, ..., exact = TRUE)

# S3 method for soma_adat
[(x, i, j, ...) <- value

# S3 method for soma_adat
$(x, i, j, ...) <- value

# S3 method for soma_adat
[[(x, i, j, ...) <- value

# S3 method for soma_adat
median(x, na.rm = FALSE, ...)

Arguments

x, object

A soma_adat class object.

show_header

Logical. Should all the Header Data information be displayed instead of the data frame (tibble) object?

...

potentially further arguments for methods; not used in the default method.

tbl

An annotations table. If NULL (default), annotation information is extracted from the object itself (if possible). Alternatively, the result of a call to getAnalyteInfo(), from which Target names can be extracted.

digits

Integer. Used for number formatting with signif().

i

indices specifying elements to extract or replace. Indices are numeric or character vectors or empty (missing) or NULL. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero). Character vectors will be matched to the names of the object (or for matrices/arrays, the dimnames): see ‘Character indices’ below for further details.

For [-indexing only: i, j, ... can be logical vectors, indicating elements/slices to select. Such vectors are recycled if necessary to match the corresponding extent. i, j, ... can also be negative integers, indicating elements/slices to leave out of the selection.

When indexing arrays by [ a single argument i can be a matrix with as many columns as there are dimensions of x; the result is then a vector with elements corresponding to the sets of indices in each row of i.

An index value of NULL is treated as if it were integer(0).

j

indices specifying elements to extract or replace. Indices are numeric or character vectors or empty (missing) or NULL. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero). Character vectors will be matched to the names of the object (or for matrices/arrays, the dimnames): see ‘Character indices’ below for further details.

For [-indexing only: i, j, ... can be logical vectors, indicating elements/slices to select. Such vectors are recycled if necessary to match the corresponding extent. i, j, ... can also be negative integers, indicating elements/slices to leave out of the selection.

When indexing arrays by [ a single argument i can be a matrix with as many columns as there are dimensions of x; the result is then a vector with elements corresponding to the sets of indices in each row of i.

An index value of NULL is treated as if it were integer(0).

drop

For matrices and arrays. If TRUE the result is coerced to the lowest possible dimension (see the examples). This only works for extracting elements, not for the replacement. See drop for further details.

name

A literal character string or a name (possibly backtick quoted). For extraction, this is normally (see under ‘Environments’) partially matched to the names of the object.

exact

Ignored with a warning().

value

typically an array-like R object of a similar class as x.

na.rm

a logical value indicating whether NA values should be stripped before the computation proceeds.

See also

Examples

my_adat <- system.file("example", "example_data.adat",
                       package = "SomaDataIO", mustWork = TRUE) |> read_adat()
# S3 print method
my_adat
#> ══ SomaScan Data ═══════════════════════════════════════════════════════════════
#>      Attributes intact    ✓
#>      Rows                 192
#>      Columns              5318
#>      Clinical Data        34
#>      Features             5284
#> ── Column Meta ─────────────────────────────────────────────────────────────────
#> ℹ SeqId, SeqIdVersion, SomaId, TargetFullName, Target, UniProt,
#> ℹ EntrezGeneID, EntrezGeneSymbol, Organism, Units, Type, Dilution,
#> ℹ PlateScale_Reference, CalReference, Cal_Example_Adat_Set001,
#> ℹ ColCheck, CalQcRatio_Example_Adat_Set001_170255, QcReference_170255,
#> ℹ Cal_Example_Adat_Set002, CalQcRatio_Example_Adat_Set002_170255,
#> ℹ Dilution2
#> ── Tibble ──────────────────────────────────────────────────────────────────────
#> # A tibble: 192 × 5,319
#>    row_names      PlateId  PlateRunDate ScannerID PlatePosition SlideId Subarray
#>    <chr>          <chr>    <chr>        <chr>     <chr>           <dbl>    <dbl>
#>  1 258495800012_3 Example… 2020-06-18   SG152144… H9            2.58e11        3
#>  2 258495800004_7 Example… 2020-06-18   SG152144… H8            2.58e11        7
#>  3 258495800010_8 Example… 2020-06-18   SG152144… H7            2.58e11        8
#>  4 258495800003_4 Example… 2020-06-18   SG152144… H6            2.58e11        4
#>  5 258495800009_4 Example… 2020-06-18   SG152144… H5            2.58e11        4
#>  6 258495800012_8 Example… 2020-06-18   SG152144… H4            2.58e11        8
#>  7 258495800001_3 Example… 2020-06-18   SG152144… H3            2.58e11        3
#>  8 258495800004_8 Example… 2020-06-18   SG152144… H2            2.58e11        8
#>  9 258495800001_8 Example… 2020-06-18   SG152144… H12           2.58e11        8
#> 10 258495800004_3 Example… 2020-06-18   SG152144… H11           2.58e11        3
#> # … with 182 more rows, and 5,312 more variables: SampleId <chr>,
#> #   SampleType <chr>, PercentDilution <int>, SampleMatrix <chr>, Barcode <lgl>,
#> #   Barcode2d <chr>, SampleName <lgl>, SampleNotes <lgl>,
#> #   AliquotingNotes <lgl>, SampleDescription <chr>, AssayNotes <lgl>,
#> #   TimePoint <lgl>, ExtIdentifier <lgl>, SsfExtId <lgl>, SampleGroup <lgl>, …
#> ════════════════════════════════════════════════════════════════════════════════

# S3 summary method
# MMP (4) analytes
mmps <- c("seq.2579.17", "seq.2788.55", "seq.2789.26", "seq.4925.54")
summary(my_adat[, mmps])
#>  seq.2579.17         seq.2788.55         seq.2789.26         
#>  Target : MMP-9      Target : MMP-3      Target : MMP-7      
#>  Min    :    59.3    Min    :  29.30     Min    :    24.1    
#>  1Q     :  9360.5    1Q     : 126.67     1Q     :  8182.0    
#>  Median : 13191.2    Median : 149.75     Median : 10947.8    
#>  Mean   : 14372.5    Mean   : 157.78     Mean   : 11277.6    
#>  3Q     : 19387.3    3Q     : 179.08     3Q     : 13577.0    
#>  Max    : 51263.0    Max    : 423.20     Max    : 33496.9    
#>  sd     :  7449.9    sd     :  53.57     sd     :  5802.6    
#>  MAD    :  6877.9    MAD    :  38.62     MAD    :  4001.2    
#>  IQR    : 10026.8    IQR    :  52.40     IQR    :  5395.0    
#>  seq.4925.54         
#>  Target : MMP-13     
#>  Min    :   25.20    
#>  1Q     :  328.30    
#>  Median :  389.70    
#>  Mean   :  424.48    
#>  3Q     :  452.85    
#>  Max    : 2883.20    
#>  sd     :  259.01    
#>  MAD    :   93.85    
#>  IQR    :  124.55    

# Summarize by group
my_adat[, mmps] |>
  split(my_adat$Sex) |>
  lapply(summary)
#> $F
#>  seq.2579.17         seq.2788.55         seq.2789.26         
#>  Target : MMP-9      Target : MMP-3      Target : MMP-7      
#>  Min    :  4484      Min    :  97.10     Min    :   289.7    
#>  1Q     :  8336      1Q     : 127.70     1Q     :  9410.0    
#>  Median : 12859      Median : 150.60     Median : 12157.2    
#>  Mean   : 13674      Mean   : 156.02     Mean   : 12877.3    
#>  3Q     : 16148      3Q     : 181.10     3Q     : 15151.9    
#>  Max    : 31879      Max    : 235.40     Max    : 32221.3    
#>  sd     :  6389      sd     :  34.66     sd     :  5214.4    
#>  MAD    :  6488      MAD    :  38.10     MAD    :  4342.7    
#>  IQR    :  7811      IQR    :  53.40     IQR    :  5741.9    
#>  seq.4925.54         
#>  Target : MMP-13     
#>  Min    :  241.80    
#>  1Q     :  333.80    
#>  Median :  388.10    
#>  Mean   :  422.14    
#>  3Q     :  460.70    
#>  Max    : 1353.80    
#>  sd     :  171.89    
#>  MAD    :   95.78    
#>  IQR    :  126.90    
#> 
#> $M
#>  seq.2579.17         seq.2788.55         seq.2789.26         
#>  Target : MMP-9      Target : MMP-3      Target : MMP-7      
#>  Min    :  5894      Min    :  99.90     Min    :  2850      
#>  1Q     :  9973      1Q     : 132.40     1Q     :  8640      
#>  Median : 13645      Median : 157.40     Median : 11230      
#>  Mean   : 15715      Mean   : 173.63     Mean   : 11699      
#>  3Q     : 20595      3Q     : 196.20     3Q     : 13061      
#>  Max    : 51263      Max    : 423.20     Max    : 33497      
#>  sd     :  7979      sd     :  61.17     sd     :  5243      
#>  MAD    :  6344      MAD    :  40.18     MAD    :  3477      
#>  IQR    : 10622      IQR    :  63.80     IQR    :  4421      
#>  seq.4925.54         
#>  Target : MMP-13     
#>  Min    :  205.7     
#>  1Q     :  329.3     
#>  Median :  401.6     
#>  Mean   :  462.7     
#>  3Q     :  468.9     
#>  Max    : 2883.2     
#>  sd     :  331.3     
#>  MAD    :  103.8     
#>  IQR    :  139.6     
#> 

# Alternatively pass annotations with Target info
anno <- getAnalyteInfo(my_adat)
summary(my_adat[, mmps], tbl = anno)
#>  seq.2579.17         seq.2788.55         seq.2789.26         
#>  Target : MMP-9      Target : MMP-3      Target : MMP-7      
#>  Min    :    59.3    Min    :  29.30     Min    :    24.1    
#>  1Q     :  9360.5    1Q     : 126.67     1Q     :  8182.0    
#>  Median : 13191.2    Median : 149.75     Median : 10947.8    
#>  Mean   : 14372.5    Mean   : 157.78     Mean   : 11277.6    
#>  3Q     : 19387.3    3Q     : 179.08     3Q     : 13577.0    
#>  Max    : 51263.0    Max    : 423.20     Max    : 33496.9    
#>  sd     :  7449.9    sd     :  53.57     sd     :  5802.6    
#>  MAD    :  6877.9    MAD    :  38.62     MAD    :  4001.2    
#>  IQR    : 10026.8    IQR    :  52.40     IQR    :  5395.0    
#>  seq.4925.54         
#>  Target : MMP-13     
#>  Min    :   25.20    
#>  1Q     :  328.30    
#>  Median :  389.70    
#>  Mean   :  424.48    
#>  3Q     :  452.85    
#>  Max    : 2883.20    
#>  sd     :  259.01    
#>  MAD    :   93.85    
#>  IQR    :  124.55